# backend/app/routers/exams.py
"""Exams router — /api/v1/exams/*"""
from __future__ import annotations

import re
import uuid
from datetime import UTC, datetime
from typing import Any
from urllib.parse import quote

from fastapi import APIRouter, Depends, HTTPException, Response, status
from sqlalchemy import delete, or_, select, update
from sqlalchemy.exc import IntegrityError
from sqlalchemy.ext.asyncio import AsyncSession

from app.database import get_db
from app.dependencies import get_current_teacher, get_exam_for_teacher
from app.models.exam import Exam
from app.models.exam_exercise import ExamExercise
from app.models.exam_mc_group import ExamMcGroup
from app.models.exercise import Exercise
from app.models.teacher import Teacher
from app.routers.exercises import parse_exercise_score
from app.schemas.exam import (
    ExamCreate,
    ExamMcGroupResponse,
    ExamResponse,
    ExamUpdate,
    ExerciseResponse,
)
from app.services.capabilities import require_server_latex
from app.services.exercise_resource_store import load_resources_for_exercises
from app.services.latex import CompilationError, compile_exam_latex
from app.services.latex_resources import ResourceError
from app.services.logo import resolve_logo

router = APIRouter(prefix="/exams", tags=["exams"])


def _to_exercise_response(
    ex: Exercise,
    order_index: int = 1,
    mc_group_id: uuid.UUID | None = None,
    sub_index: int | None = None,
) -> ExerciseResponse:
    return ExerciseResponse(
        id=ex.id,
        teacher_id=ex.teacher_id,
        name=ex.name,
        topic_tag=ex.topic_tag,
        grade=ex.grade,
        subject=ex.subject,
        latex_body=ex.latex_body,
        code_withheld=ex.code_withheld,
        max_points=ex.max_points,
        version=ex.version,
        exercise_group_id=ex.exercise_group_id,
        variant_key=ex.variant_key,
        is_current=ex.is_current,
        order_index=order_index,
        question_type=ex.question_type,
        correct_answers=ex.correct_answers,
        penalty=ex.penalty,
        mc_group_id=mc_group_id,
        sub_index=sub_index,
    )


ExerciseRow = tuple[Exercise, int, uuid.UUID | None, int | None]


def _to_exam_response(
    e: Exam,
    exercises: list[ExerciseRow],
    mc_groups: list[ExamMcGroup],
) -> ExamResponse:
    group_members: dict[uuid.UUID, list[tuple[int, uuid.UUID]]] = {}
    for ex, _, gid, sidx in exercises:
        if gid is not None:
            group_members.setdefault(gid, []).append((sidx or 0, ex.id))

    return ExamResponse(
        id=e.id,
        teacher_id=e.teacher_id,
        title=e.title,
        latex_template=e.latex_template,
        compilation_status=e.compilation_status,
        created_at=e.created_at,
        retention_until=e.retention_until,
        testart=e.testart,
        grade=e.grade,
        klasse=e.klasse,
        datum=e.datum,
        nr=e.nr,
        fach=e.fach,
        lehrernachname=e.lehrernachname,
        info_text=e.info_text,
        grading_key=e.grading_key,
        exercises=[_to_exercise_response(ex, idx, gid, sidx) for ex, idx, gid, sidx in exercises],
        mc_groups=[
            ExamMcGroupResponse(
                id=g.id,
                exam_id=g.exam_id,
                title=g.title,
                scoring_text=g.scoring_text,
                order_index=g.order_index,
                member_ids=[
                    mid for _, mid in sorted(group_members.get(g.id, []))
                ],
            )
            for g in mc_groups
        ],
    )


async def _fetch_exam_exercises(exam_id: uuid.UUID, db: AsyncSession) -> list[ExerciseRow]:
    return (await _fetch_exercises_for_exams([exam_id], db))[exam_id]


async def _resolve_linkable_exercise(
    exercise_id: uuid.UUID, teacher: Teacher, db: AsyncSession
) -> Exercise:
    """
    Load an exercise *teacher* is allowed to place into their own exam.

    Own exercises plus explicitly published ones. Unknown or foreign private
    exercises raise 404 — without this check, linking an arbitrary UUID turns
    the exam read/compile endpoints into a cross-tenant disclosure channel.
    """
    res = await db.execute(
        select(Exercise).where(
            Exercise.id == exercise_id,
            or_(Exercise.teacher_id == teacher.id, Exercise.is_public.is_(True)),
        )
    )
    ex = res.scalar_one_or_none()
    if ex is None:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail="Referenced exercise not found.",
        )
    return ex


def _resolve_mc_group_id(
    client_group_id: uuid.UUID | None,
    id_map: dict[uuid.UUID, uuid.UUID],
) -> uuid.UUID | None:
    """
    Map a client-supplied MC group id onto the group just persisted for this exam.

    An id that is not in *id_map* does not belong to this exam's groups and is
    rejected rather than stored verbatim.
    """
    if client_group_id is None:
        return None
    resolved = id_map.get(client_group_id)
    if resolved is None:
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST,
            detail="Unknown mc_group_id for this exam.",
        )
    return resolved


def _dedupe_exercise_links(links: list[Any]) -> list[Any]:
    """
    Collapse repeated exercise_ids in an exercise_links payload.

    (exam_id, exercise_id) is the exam_exercises primary key, so an exercise
    listed twice — e.g. once standalone and once as an MC group member — used to
    raise an IntegrityError and surface as a 500. The entry carrying MC group
    membership wins; otherwise the first occurrence does.
    """
    kept: dict[uuid.UUID, Any] = {}
    order: list[uuid.UUID] = []
    for link in links:
        existing = kept.get(link.exercise_id)
        if existing is None:
            kept[link.exercise_id] = link
            order.append(link.exercise_id)
        elif existing.mc_group_id is None and link.mc_group_id is not None:
            kept[link.exercise_id] = link
    return [kept[eid] for eid in order]


async def _fetch_exam_mc_groups(exam_id: uuid.UUID, db: AsyncSession) -> list[ExamMcGroup]:
    return (await _fetch_mc_groups_for_exams([exam_id], db))[exam_id]


async def _fetch_exercises_for_exams(
    exam_ids: list[uuid.UUID], db: AsyncSession
) -> dict[uuid.UUID, list[ExerciseRow]]:
    """Every listed exam's exercises in one query (plus one for legacy links)."""
    by_exam: dict[uuid.UUID, list[ExerciseRow]] = {exam_id: [] for exam_id in exam_ids}
    if not exam_ids:
        return by_exam

    result = await db.execute(
        select(
            ExamExercise.exam_id,
            Exercise,
            ExamExercise.order_index,
            ExamExercise.mc_group_id,
            ExamExercise.sub_index,
        )
        .join(ExamExercise, Exercise.id == ExamExercise.exercise_id)
        .where(ExamExercise.exam_id.in_(exam_ids))
        .order_by(ExamExercise.order_index.asc(), ExamExercise.sub_index.asc())
    )
    for exam_id, exercise, order, group_id, sub_index in result.all():
        by_exam[exam_id].append((exercise, order, group_id, sub_index))

    # Legacy exams link via exercise.exam_id instead of the junction table.
    legacy_ids = [exam_id for exam_id, rows in by_exam.items() if not rows]
    if legacy_ids:
        legacy = await db.execute(
            select(Exercise)
            .where(Exercise.exam_id.in_(legacy_ids))
            .order_by(Exercise.order_index.asc())
        )
        for exercise in legacy.scalars().all():
            if exercise.exam_id is not None:
                by_exam[exercise.exam_id].append((exercise, exercise.order_index, None, None))

    return by_exam


async def _fetch_mc_groups_for_exams(
    exam_ids: list[uuid.UUID], db: AsyncSession
) -> dict[uuid.UUID, list[ExamMcGroup]]:
    """Every listed exam's MC groups in one query, keyed by exam id."""
    by_exam: dict[uuid.UUID, list[ExamMcGroup]] = {exam_id: [] for exam_id in exam_ids}
    if not exam_ids:
        return by_exam
    result = await db.execute(
        select(ExamMcGroup)
        .where(ExamMcGroup.exam_id.in_(exam_ids))
        .order_by(ExamMcGroup.order_index.asc())
    )
    for group in result.scalars().all():
        by_exam[group.exam_id].append(group)
    return by_exam


@router.get("", response_model=list[ExamResponse])
async def list_exams(
    grade: str | None = None,
    subject: str | None = None,
    search: str | None = None,
    teacher: Teacher = Depends(get_current_teacher),
    db: AsyncSession = Depends(get_db),
) -> list[ExamResponse]:
    """List all non-deleted exams belonging to the authenticated teacher."""
    query = select(Exam).where(
        Exam.teacher_id == teacher.id,
        Exam.deleted_at.is_(None),
    )

    if grade:
        query = query.where(Exam.grade == grade)

    if subject:
        query = query.where(Exam.fach == subject)

    if search:
        search_pattern = f"%{search}%"
        query = query.where(
            or_(
                Exam.title.ilike(search_pattern),
                Exam.grade.ilike(search_pattern),
                Exam.klasse.ilike(search_pattern),
                Exam.fach.ilike(search_pattern),
            )
        )

    query = query.order_by(Exam.created_at.desc())
    result = await db.execute(query)
    exams = list(result.scalars().all())

    exam_ids = [e.id for e in exams]
    exercises_by_exam = await _fetch_exercises_for_exams(exam_ids, db)
    groups_by_exam = await _fetch_mc_groups_for_exams(exam_ids, db)

    return [_to_exam_response(e, exercises_by_exam[e.id], groups_by_exam[e.id]) for e in exams]


async def _persist_mc_groups(
    exam_id: uuid.UUID,
    mc_groups_data: list[Any],
    db: AsyncSession,
) -> dict[uuid.UUID, uuid.UUID]:
    """
    Create ExamMcGroup rows and return the {client_id: db_id} mapping.

    A client-supplied id is kept verbatim, exactly like Exam.id and Exercise.id.
    Minting a fresh id here instead meant every save handed the client a *new*
    group id, so the local (IndexedDB) group records and everything referencing
    them drifted out of sync with the server on every single write.
    """
    id_map: dict[uuid.UUID, uuid.UUID] = {}
    for g in mc_groups_data:
        group_id = g.id or uuid.uuid4()
        group = ExamMcGroup(
            id=group_id,
            exam_id=exam_id,
            title=g.title,
            scoring_text=g.scoring_text,
            order_index=g.order_index,
        )
        db.add(group)
        # Both directions resolve, so a payload may address a group either by the
        # id it sent or by the id a previous response handed back.
        id_map[group_id] = group_id
        if g.id is not None:
            id_map[g.id] = group_id
    try:
        await db.flush()
    except IntegrityError:
        await db.rollback()
        raise HTTPException(
            status_code=status.HTTP_409_CONFLICT,
            detail="An MC group with this id already exists.",
        ) from None
    return id_map


@router.post("", response_model=ExamResponse, status_code=status.HTTP_201_CREATED)
async def create_exam(
    body: ExamCreate,
    teacher: Teacher = Depends(get_current_teacher),
    db: AsyncSession = Depends(get_db),
) -> ExamResponse:
    """Create a new exam and live-link exercises from library or inline creation."""
    kwargs = {
        "teacher_id": teacher.id,
        "title": body.title,
        "latex_template": body.latex_template,
        "retention_until": body.retention_until,
        "testart": body.testart,
        "grade": body.grade,
        "klasse": body.klasse,
        "datum": body.datum,
        "nr": body.nr,
        "fach": body.fach,
        "lehrernachname": body.lehrernachname,
        "info_text": body.info_text,
        "grading_key": body.grading_key,
    }
    if body.id:
        kwargs["id"] = body.id
    exam = Exam(**kwargs)
    db.add(exam)
    # Client-chosen ids let local-mode records keep their UUID when syncing. A
    # collision must be a deterministic 409, not an IntegrityError surfacing as
    # a 500 — which would distinguish "id exists" from "id is free".
    try:
        await db.flush()
    except IntegrityError:
        await db.rollback()
        raise HTTPException(
            status_code=status.HTTP_409_CONFLICT,
            detail="An exam with this id already exists.",
        ) from None

    # Persist MC groups first so we can map client IDs to DB IDs
    mc_group_id_map: dict[uuid.UUID, uuid.UUID] = {}
    if body.mc_groups:
        mc_group_id_map = await _persist_mc_groups(exam.id, body.mc_groups, db)

    linked_exercises: list[ExerciseRow] = []
    order = 1

    # 1. Link existing library exercises via exercise_links (mc-aware)
    if body.exercise_links:
        for link_data in _dedupe_exercise_links(body.exercise_links):
            ex = await _resolve_linkable_exercise(link_data.exercise_id, teacher, db)
            resolved_group_id = _resolve_mc_group_id(link_data.mc_group_id, mc_group_id_map)
            effective_order = link_data.order_index or order
            link = ExamExercise(
                exam_id=exam.id,
                exercise_id=ex.id,
                order_index=effective_order,
                mc_group_id=resolved_group_id,
                sub_index=link_data.sub_index,
            )
            db.add(link)
            linked_exercises.append((ex, effective_order, resolved_group_id, link_data.sub_index))
            order = effective_order + 1

    # 2. Link existing library exercise IDs (legacy, no mc info)
    elif body.exercise_ids:
        for eid in body.exercise_ids:
            ex = await _resolve_linkable_exercise(eid, teacher, db)
            link = ExamExercise(exam_id=exam.id, exercise_id=ex.id, order_index=order)
            db.add(link)
            linked_exercises.append((ex, order, None, None))
            order += 1

    # 3. Create and link inline exercises
    if body.exercises:
        for ex_data in body.exercises:
            computed_score = (
                parse_exercise_score(ex_data.latex_body)
                if ex_data.latex_body
                else ex_data.max_points
            )
            ex = Exercise(
                teacher_id=teacher.id,
                name=ex_data.name,
                topic_tag=ex_data.topic_tag,
                latex_body=None if ex_data.code_withheld else ex_data.latex_body,
                code_withheld=ex_data.code_withheld,
                max_points=computed_score,
                version=1,
                question_type=ex_data.question_type,
                correct_answers=ex_data.correct_answers,
                penalty=ex_data.penalty,
            )
            db.add(ex)
            await db.flush()

            resolved_group_id = _resolve_mc_group_id(ex_data.mc_group_id, mc_group_id_map)

            link = ExamExercise(
                exam_id=exam.id,
                exercise_id=ex.id,
                order_index=order,
                mc_group_id=resolved_group_id,
                sub_index=ex_data.sub_index,
            )
            db.add(link)
            linked_exercises.append((ex, order, resolved_group_id, ex_data.sub_index))
            order += 1

    await db.flush()
    mc_groups = await _fetch_exam_mc_groups(exam.id, db)
    return _to_exam_response(exam, linked_exercises, mc_groups)


@router.get("/{exam_id}", response_model=ExamResponse)
async def get_exam(
    exam: Exam = Depends(get_exam_for_teacher),
    db: AsyncSession = Depends(get_db),
) -> ExamResponse:
    """Get exam details with live-linked exercises."""
    exs = await _fetch_exam_exercises(exam.id, db)
    groups = await _fetch_exam_mc_groups(exam.id, db)
    return _to_exam_response(exam, exs, groups)


@router.get("/{exam_id}/exercises", response_model=list[ExerciseResponse])
async def list_exam_exercises(
    exam: Exam = Depends(get_exam_for_teacher),
    db: AsyncSession = Depends(get_db),
) -> list[ExerciseResponse]:
    """Get exercises linked to a specific exam."""
    exs = await _fetch_exam_exercises(exam.id, db)
    return [_to_exercise_response(ex, idx, gid, sidx) for ex, idx, gid, sidx in exs]



@router.patch("/{exam_id}", response_model=ExamResponse)
async def update_exam(
    body: ExamUpdate,
    exam: Exam = Depends(get_exam_for_teacher),
    teacher: Teacher = Depends(get_current_teacher),
    db: AsyncSession = Depends(get_db),
) -> ExamResponse:
    """Update exam details and exercise links."""
    if body.title is not None:
        exam.title = body.title
    if body.latex_template is not None:
        exam.latex_template = body.latex_template
    if body.retention_until is not None:
        exam.retention_until = body.retention_until
    if body.testart is not None:
        exam.testart = body.testart
    if body.grade is not None:
        exam.grade = body.grade
    if body.klasse is not None:
        exam.klasse = body.klasse
    if body.datum is not None:
        exam.datum = body.datum
    if body.nr is not None:
        exam.nr = body.nr
    if body.fach is not None:
        exam.fach = body.fach
    if body.lehrernachname is not None:
        exam.lehrernachname = body.lehrernachname
    if body.info_text is not None:
        exam.info_text = body.info_text
    if body.grading_key is not None:
        exam.grading_key = body.grading_key

    if body.mc_groups is not None:
        # Detach the links before dropping the groups. exam_exercises.mc_group_id
        # is a real FK, so deleting a group would otherwise take its member rows
        # with it — silently removing those exercises from the exam.
        await db.execute(
            update(ExamExercise)
            .where(ExamExercise.exam_id == exam.id)
            .values(mc_group_id=None, sub_index=None)
        )
        await db.execute(delete(ExamMcGroup).where(ExamMcGroup.exam_id == exam.id))
        await db.flush()
        mc_group_id_map = await _persist_mc_groups(exam.id, body.mc_groups, db)
    else:
        # Groups untouched: links may still address the groups already on the exam.
        existing_groups = await _fetch_exam_mc_groups(exam.id, db)
        mc_group_id_map = {g.id: g.id for g in existing_groups}

    if body.exercise_links is not None:
        await db.execute(delete(ExamExercise).where(ExamExercise.exam_id == exam.id))
        await db.flush()
        for link_data in _dedupe_exercise_links(body.exercise_links):
            ex = await _resolve_linkable_exercise(link_data.exercise_id, teacher, db)
            resolved_group_id = _resolve_mc_group_id(link_data.mc_group_id, mc_group_id_map)
            db.add(ExamExercise(
                exam_id=exam.id,
                exercise_id=ex.id,
                order_index=link_data.order_index,
                mc_group_id=resolved_group_id,
                sub_index=link_data.sub_index,
            ))
    elif body.exercise_ids is not None:
        await db.execute(delete(ExamExercise).where(ExamExercise.exam_id == exam.id))
        order = 1
        for eid in body.exercise_ids:
            ex = await _resolve_linkable_exercise(eid, teacher, db)
            db.add(ExamExercise(exam_id=exam.id, exercise_id=ex.id, order_index=order))
            order += 1

    await db.flush()
    exs = await _fetch_exam_exercises(exam.id, db)
    groups = await _fetch_exam_mc_groups(exam.id, db)
    return _to_exam_response(exam, exs, groups)


@router.delete("/{exam_id}", status_code=status.HTTP_204_NO_CONTENT)
async def delete_exam(
    exam: Exam = Depends(get_exam_for_teacher),
    db: AsyncSession = Depends(get_db),
) -> None:
    """Soft-delete an exam."""
    exam.deleted_at = datetime.now(tz=UTC)


@router.post("/{exam_id}/compile")
async def compile_exam_endpoint(
    answers: bool = False,
    exam: Exam = Depends(get_exam_for_teacher),
    teacher: Teacher = Depends(require_server_latex),
    db: AsyncSession = Depends(get_db),
) -> Response:
    """Compile complete exam LaTeX document using live-linked library exercises."""
    ex_tuples = await _fetch_exam_exercises(exam.id, db)
    mc_groups = await _fetch_exam_mc_groups(exam.id, db)

    # Resources live on the server in server/hybrid mode, so the browser never
    # uploads them for a compile — they are read straight out of the database.
    try:
        binary_files = await load_resources_for_exercises(
            [t[0].id for t in ex_tuples if getattr(t[0], "id", None)], teacher.id, db
        )
    except ResourceError as exc:
        raise HTTPException(
            status_code=status.HTTP_422_UNPROCESSABLE_CONTENT,
            detail=str(exc),
            headers={"code": "ERR_COMPILE_FAILED"},
        ) from exc

    _, logo = await resolve_logo(teacher.id, exam.id, db)

    try:
        pdf_bytes = await compile_exam_latex(
            exam,
            ex_tuples,
            mc_groups,
            show_answers=answers,
            binary_files=binary_files,
            logo=logo,
        )
        exam.compilation_status = "compiled"
        await db.flush()
    except TimeoutError:
        exam.compilation_status = "failed"
        await db.flush()
        raise HTTPException(
            status_code=status.HTTP_504_GATEWAY_TIMEOUT,
            detail="Compilation timed out.",
            headers={"code": "ERR_COMPILE_TIMEOUT"},
        ) from None
    except CompilationError as exc:
        exam.compilation_status = "failed"
        await db.flush()
        raise HTTPException(
            status_code=status.HTTP_422_UNPROCESSABLE_CONTENT,
            detail=str(exc),
            headers={"code": "ERR_COMPILE_FAILED"},
        ) from exc
    except OSError as exc:
        exam.compilation_status = "failed"
        await db.flush()
        raise HTTPException(
            status_code=status.HTTP_503_SERVICE_UNAVAILABLE,
            detail="The LaTeX compilation service is unavailable.",
            headers={"code": "ERR_COMPILE_UNAVAILABLE"},
        ) from exc

    # RFC 6266. The title is user-controlled: a quote would break out of the
    # quoted-string and let the download filename be spoofed, and a CR/LF would
    # be rejected downstream as a malformed header.
    safe_title = re.sub(r"[^A-Za-z0-9._-]", "_", exam.title)[:100] or "exam"
    filename = f"Exam_{safe_title}{'_answers' if answers else ''}.pdf"
    return Response(
        content=pdf_bytes,
        media_type="application/pdf",
        headers={
            "Content-Disposition": (
                f'attachment; filename="{filename}"; '
                f"filename*=UTF-8''{quote(filename)}"
            )
        },
    )